Fix log output if command fails

+++ .bin/functions.sh
@@ -153,16 +153,19 @@ logOutputFromBackgroundProcesses() {
         title="${PID_LOG_TITLE[$pid]}"
         log="${PID_LOG_FILE[$pid]}"
 
-        echo "$title"
+        if [[ -s "$log" ]]; then
+            echo "-- $title --:"
 
-        cat "$log"
-        rm -f -- "$log"
+            cat "$log"
+            rm -f -- "$log"
+
+            echo ""
+            echo ""
+        fi
 
         unset PID_LOG_TITLE[$pid]
         unset PID_LOG_FILE[$pid]
 
-        echo ""
-        echo ""
     done
 }
